# api/openapi/pocketcoder.yaml
openapi: 3.0.3
info:
  title: PocketCoder operation API
  version: 1.0.0
  description: Spec-first contract for PocketCoder custom operations. PocketBase collection APIs are out of scope.
servers:
  - url: /
paths:
  /api/pocketcoder/v1/chats/{chatId}/session/prompt:
    post:
      operationId: promptChat
      tags: [agent]
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/PromptRequest' } } } }
      responses: { '202': { description: Run accepted, content: { application/json: { schema: { $ref: '#/components/schemas/AcceptedResponse' } } } }, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '503': {$ref: '#/components/responses/Unavailable'} }
  /api/pocketcoder/v1/chats/{chatId}/stream:
    get:
      operationId: streamChatEvents
      tags: [agent]
      x-pocketcoder-transport: sse
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }, { $ref: '#/components/parameters/Cursor' }]
      responses: { '200': { description: Incremental AG-UI events, content: { text/event-stream: { schema: { $ref: '#/components/schemas/AgentEvent' } } } }, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'} }
  /api/pocketcoder/v1/chats/{chatId}/session/cancel:
    post:
      operationId: cancelChatSession
      tags: [agent]
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }]
      responses: { '202': {description: Cancellation accepted}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'} }
  /api/pocketcoder/v1/chats/{chatId}/session/set-mode:
    post:
      operationId: setChatMode
      tags: [agent]
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ModeRequest'}}}}
      responses: {'202': {description: Operation accepted}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '502': {$ref: '#/components/responses/BadGateway'}, '503': {$ref: '#/components/responses/Unavailable'} }
  /api/pocketcoder/v1/chats/{chatId}/session/set-config-option:
    post:
      operationId: setChatConfigOption
      tags: [agent]
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ConfigOptionRequest'}}}}
      responses: {'202': {description: Operation accepted}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '502': {$ref: '#/components/responses/BadGateway'}, '503': {$ref: '#/components/responses/Unavailable'} }
  /api/pocketcoder/v1/chats/{chatId}/session/request-permission/{id}:
    post:
      operationId: respondToPermission
      tags: [agent]
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DecisionRequest'}}}}
      responses: {'202': {description: Operation accepted}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '502': {$ref: '#/components/responses/BadGateway'}, '503': {$ref: '#/components/responses/Unavailable'} }
  /api/pocketcoder/v1/chats/{chatId}/session/elicitation/{id}:
    post:
      operationId: respondToElicitation
      tags: [agent]
      security: [{pocketbaseToken: []}]
      parameters: [{ $ref: '#/components/parameters/ChatId' }, { $ref: '#/components/parameters/RequestId' }]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/DecisionRequest'}}}}
      responses: {'202': {description: Operation accepted}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '502': {$ref: '#/components/responses/BadGateway'}, '503': {$ref: '#/components/responses/Unavailable'} }
...
  /api/pocketcoder/v1/push:
    post:
      operationId: sendPushNotification
      tags: [push]
      security: [{pocketbaseToken: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PushRequest'}}}}
      responses: {'204': {description: Operation completed}, '200': {$ref: '#/components/responses/JsonSuccess'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'} }
...
components:
  securitySchemes:
    pocketbaseToken: {type: apiKey, in: header, name: Authorization, description: Raw PocketBase token; do not prefix with Bearer.}
  parameters:
    ChatId: {name: chatId, in: path, required: true, schema: {type: string, minLength: 1}}
    RequestId: {name: id, in: path, required: true, schema: {type: string, minLength: 1}}
    Cursor: {name: cursor, in: query, required: false, schema: {type: integer, minimum: 0}}
...
  responses:
    JsonSuccess: {description: Successful JSON response, content: {application/json: {schema: {$ref: '#/components/schemas/JsonObject'}}}}
    BadRequest: {description: Invalid request, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    Unauthorized: {description: Authentication required, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
    NotFound: {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
    Unavailable: {description: Service unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    BadGateway: {description: Upstream service unavailable, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  schemas:
    JsonObject: {type: object, additionalProperties: true}
    ErrorResponse: {type: object, required: [code, message, data], additionalProperties: false, properties: {code: {type: integer}, message: {type: string}, data: {type: object, additionalProperties: true}}}
    AcceptedResponse: {type: object, required: [runId], additionalProperties: false, properties: {runId: {type: string, minLength: 1}}}
...
    PromptRequest: {type: object, required: [prompt], additionalProperties: false, properties: {prompt: {type: array, minItems: 1, items: {$ref: '#/components/schemas/ContentBlock'}}}}
    ContentBlock: {type: object, required: [type], additionalProperties: true, properties: {type: {type: string, minLength: 1}, text: {type: string}}}
    ModeRequest: {type: object, required: [modeId], additionalProperties: false, properties: {modeId: {type: string, minLength: 1}}}
    ConfigOptionRequest: {type: object, required: [configId, value], additionalProperties: false, properties: {configId: {type: string, minLength: 1}, value: {type: string}}}
    DecisionRequest: {type: object, additionalProperties: true}
...
    PushRequest: {type: object, additionalProperties: true}
...
    AgentEvent: {type: object, additionalProperties: true}
...